requirements.txt
# Usage: pip install -r requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil 
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
torch>=1.7.0 
torchvision>=0.8.1
tqdm>=4.64.0
fastapi>=0.87.0
uvicorn>=0.19.0
pymongo>=4.3.3
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
python-multipart>=0.0.5
